commsgaze:Fixed camera locked first time init issue
