+++ ProjectSpace/commsgaze/src/main/java/com/pwc/commsgaze/MainViewModel.java
@@ -19,7 +19,7 @@ import java.util.Arrays;
 /*Holds details for MainActivity in a Activity life-conscious way */
 public class MainViewModel extends ViewModel {
 
-    private MutableLiveData<Boolean> isFirstRun;
+
     private static final String TAG = "MainViewModel";
     private DetectionEngineMaker detectionEngineMakerInstance;
     private Detector detector;
@@ -158,18 +158,5 @@ public class MainViewModel extends ViewModel {
     }
 
 
-    /*Check on UI thread for shared preference before calling this*/
-    LiveData<Boolean> getIsFirstRun() {
-        if (isFirstRun == null) {
-            isFirstRun = new MutableLiveData<Boolean>();
-            isFirstRun.setValue(true);
-            Log.d(TAG,"getIsFirstRunCalled");
-        }
-        return isFirstRun;
-    }
 
-    public void initialisationDone() {
-        Log.d(TAG," InitialisationDone Called");
-        isFirstRun.setValue(false);
-    }
 }
