+++ ProjectSpace/commsgaze/src/main/res/layout/fragment_initialisation.xml
@@ -1,13 +1,13 @@
 <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
+    android:layout_height="wrap_content"
     tools:context="com.pwc.commsgaze.initialisation.InitialisationFragment">
 
     <LinearLayout
         android:id="@+id/initialisationLinearLayout"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
+        android:layout_height="match_parent"
         android:orientation="vertical"
         android:layout_gravity="center"
         android:padding="50dp">
